+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change Wolf armor helmet drake trophy cost.
+
 ## [0.6.25] - 2024-06-18
 
 ### Changed