+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.6.26] - 2024-06-19
+
 ### Changed
 
 - Rebalance shields and tower shields movement speed and parry bonus.
@@ -308,7 +310,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial release.
 
-[unreleased]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.25...HEAD
+[unreleased]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.26...HEAD
+[0.6.26]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.25...0.6.26
 [0.6.25]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.24...0.6.25
 [0.6.24]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.23...0.6.24
 [0.6.23]: https://github.com/nbusseneau/LotusEcarlateChanges/compare/0.6.22...0.6.23